🌹 split open_humans.views into separate files
