+++ open_humans/views.py
@@ -1,28 +1,17 @@
 import re
 import urlparse
 
-from operator import attrgetter
-
-from account.views import (LoginView as AccountLoginView,
-                           SettingsView as AccountSettingsView,
-                           SignupView as AccountSignupView)
-
 from django.apps import apps
 from django.conf import settings
-from django.contrib import messages as django_messages
-from django.contrib.auth import logout
 from django.core.urlresolvers import reverse, reverse_lazy
 from django.db.models import Count
-from django.http import Http404, HttpResponse, HttpResponseRedirect
+from django.http import HttpResponseRedirect
 from django.shortcuts import redirect
-from django.views.generic.base import RedirectView, TemplateView, View
-from django.views.generic.detail import DetailView
-from django.views.generic.edit import DeleteView, FormView, UpdateView
-from django.views.generic.list import ListView
+from django.views.generic.base import TemplateView, View
+from django.views.generic.edit import DeleteView
 
 from oauth2_provider.models import (
-    get_application_model as get_oauth2_application_model,
-    AccessToken)
+    get_application_model as get_oauth2_application_model)
 from oauth2_provider.views.base import (
     AuthorizationView as OriginalAuthorizationView)
 from oauth2_provider.exceptions import OAuthToolkitError
@@ -31,292 +20,11 @@ from common.mixins import NeverCacheMixin, PrivateMixin
 from common.utils import querydict_from_dict, get_source_labels
 
 from activities.data_selfie.models import DataSelfieDataFile
-from activities.runkeeper.models import UserData as UserDataRunKeeper
-from activities.twenty_three_and_me.models import (
-    UserData as UserDataTwentyThreeAndMe)
-from data_import.models import DataFile, DataRetrievalTask
+from data_import.models import DataFile
 from public_data.models import PublicDataAccess
-from studies.models import StudyGrant
-from studies.american_gut.models import UserData as UserDataAmericanGut
-from studies.go_viral.models import UserData as UserDataGoViral
-from studies.pgp.models import UserData as UserDataPgp
-from studies.wildlife.models import UserData as UserDataWildLife
 
-from .forms import (EmailUserForm,
-                    MemberLoginForm,
-                    MemberSignupForm,
-                    MyMemberChangeEmailForm,
-                    MyMemberChangeNameForm,
-                    MyMemberContactSettingsEditForm,
-                    MyMemberDataSelfieUpdateViewForm,
-                    MyMemberProfileEditForm)
-from .models import Member, EmailMetadata
-
-
-class MemberDetailView(DetailView):
-    """
-    Creates a view of a member's public profile.
-    """
-    queryset = Member.enriched.all()
-    template_name = 'member/member-detail.html'
-    slug_field = 'user__username'
-
-    def get_context_data(self, **kwargs):
-        """
-        Add context so login and signup return to this page.
-
-        Returning to the page is desired because the user needs to be logged in
-        to message a member from their detail page.
-        """
-        context = super(MemberDetailView, self).get_context_data(**kwargs)
-
-        context.update({
-            'next': reverse_lazy('member-detail',
-                                 kwargs={'slug': self.object.user.username}),
-        })
-
-        return context
-
-
-class MemberListView(ListView):
-    """
-    Creates a view listing members.
-    """
-    context_object_name = 'members'
-    paginate_by = 50
-    template_name = 'member/member-list.html'
-
-    def get_queryset(self):
-        if self.request.GET.get('sort') == 'username':
-            return (Member.enriched
-                    .exclude(user__username='api-administrator')
-                    .order_by('user__username'))
-
-        # First sort by name and username
-        sorted_members = sorted(Member.enriched
-                                .exclude(user__username='api-administrator'),
-                                key=attrgetter('user.username'))
-
-        # Then sort by number of badges
-        sorted_members = sorted(sorted_members,
-                                key=lambda m: len(m.badges),
-                                reverse=True)
-
-        return sorted_members
-
-    def get_context_data(self, **kwargs):
-        """
-        Add context for sorting button.
-        """
-        context = super(MemberListView, self).get_context_data(**kwargs)
-
-        if self.request.GET.get('sort') == 'username':
-            sort_direction = 'connections'
-            sort_description = 'by number of connections'
-        else:
-            sort_direction = 'username'
-            sort_description = 'by username'
-
-        context.update({
-            'sort_direction': sort_direction,
-            'sort_description': sort_description,
-        })
-
-        return context
-
-
-class MyMemberDashboardView(PrivateMixin, DetailView):
-    """
-    Creates a dashboard for the current user/member.
-
-    The dashboard also displays their public member profile.
-    """
-    context_object_name = 'member'
-    queryset = Member.enriched.all()
-    template_name = 'member/my-member-dashboard.html'
-
-    def get_object(self, queryset=None):
-        return Member.enriched.get(user=self.request.user)
-
-
-class MyMemberProfileEditView(PrivateMixin, UpdateView):
-    """
-    Creates an edit view of the current user's public member profile.
-    """
-    form_class = MyMemberProfileEditForm
-    model = Member
-    template_name = 'member/my-member-profile-edit.html'
-    success_url = reverse_lazy('my-member-dashboard')
-
-    def get_object(self, queryset=None):
-        return self.request.user.member
-
-
-class MyMemberSettingsEditView(PrivateMixin, UpdateView):
-    """
-    Creates an edit view of the current user's member account settings.
-    """
-    form_class = MyMemberContactSettingsEditForm
-    model = Member
-    template_name = 'member/my-member-settings.html'
-    success_url = reverse_lazy('my-member-settings')
-
-    def get_object(self, queryset=None):
-        return self.request.user.member
-
-
-class MyMemberChangeEmailView(PrivateMixin, AccountSettingsView):
-    """
-    Creates a view for the current user to change their email.
-
-    This is an email-only subclass of account's SettingsView.
-    """
-    form_class = MyMemberChangeEmailForm
-    template_name = 'member/my-member-change-email.html'
-    success_url = reverse_lazy('my-member-settings')
-    messages = {
-        'settings_updated': {
-            'level': django_messages.SUCCESS,
-            'text': 'Email address updated and confirmation email sent.'
-        },
-    }
-
-    def get_success_url(self, *args, **kwargs):
-        kwargs.update(
-            {'fallback_url': reverse_lazy('my-member-settings')})
-        return super(MyMemberChangeEmailView, self).get_success_url(
-            *args, **kwargs)
-
-
-class MyMemberChangeNameView(PrivateMixin, UpdateView):
-    """
-    Creates an edit view of the current member's name.
-    """
-    form_class = MyMemberChangeNameForm
-    model = Member
-    template_name = 'member/my-member-change-name.html'
-    success_url = reverse_lazy('my-member-settings')
-
-    def get_object(self, queryset=None):
-        return self.request.user.member
-
-
-class MyMemberSendConfirmationEmailView(PrivateMixin, RedirectView):
-    """
-    Send a confirmation email and redirect back to the settings page.
-    """
-    permanent = False
-    url = reverse_lazy('my-member-settings')
-
-    def get_redirect_url(self, *args, **kwargs):
-        redirect_field_name = self.request.GET.get('redirect_field_name',
-                                                   'next')
-        next_url = self.request.GET.get(redirect_field_name, self.url)
-        return next_url
-
-    def dispatch(self, request, *args, **kwargs):
-        email_address = request.user.emailaddress_set.get(primary=True)
-        email_address.send_confirmation()
-        django_messages.success(request,
-                                ('A confirmation email was sent to "{}".'
-                                 .format(email_address.email)))
-        return super(MyMemberSendConfirmationEmailView, self).dispatch(
-            request, *args, **kwargs)
-
-
-class MyMemberDatasetsView(PrivateMixin, ListView):
-    """
-    Creates a view for displaying and importing research/activity datasets.
-    """
-    template_name = 'member/my-member-research-data.html'
-    context_object_name = 'data_retrieval_tasks'
-
-    def get_queryset(self):
-        # pylint: disable=attribute-defined-outside-init
-        datasets = (DataRetrievalTask.objects
-                    .for_user(self.request.user)
-                    .grouped_recent())
-
-        return datasets
-
-    def get_context_data(self, **kwargs):
-        """
-        Add DataRetrievalTask to the request context.
-        """
-        context = super(MyMemberDatasetsView, self).get_context_data(**kwargs)
-
-        data_selfie_files = (DataFile.objects.filter(user=self.request.user,
-                                                     source='data_selfie'))
-
-        context['DataRetrievalTask'] = DataRetrievalTask
-        context['data_selfie_files'] = data_selfie_files
-
-        return context
-
-
-class MyMemberDataSelfieView(PrivateMixin, ListView):
-    """
-    Creates a view for displaying data selfie files.
-    """
-    template_name = 'member/my-member-data-selfie.html'
-    context_object_name = 'data_files'
-
-    def get_queryset(self):
-        return DataSelfieDataFile.objects.filter(user=self.request.user)
-
-
-class MyMemberDataSelfieAcknowledgeView(PrivateMixin, View):
-    """
-    Let the user acknowledge that they've seen the data selfie modal.
-    """
-    @staticmethod
-    def post(request):
-        user_data = request.user.data_selfie
-        user_data.seen_page = True
-        user_data.save()
-
-        return HttpResponse('')
-
-
-class MyMemberDataSelfieUpdateView(PrivateMixin, UpdateView):
-    """
-    Creates a view for displaying data selfie files.
-    """
-    form_class = MyMemberDataSelfieUpdateViewForm
-    model = DataFile
-    template_name = 'member/my-member-data-selfie-edit.html'
-    success_url = reverse_lazy('my-member-data-selfie')
-
-    def get_object(self, queryset=None):
-        return (DataSelfieDataFile.objects.get(id=self.kwargs['data_file'],
-                                               user=self.request.user))
-
-
-class MyMemberConnectionsView(PrivateMixin, TemplateView):
-    """
-    A view for a member to manage their connections.
-    """
-
-    template_name = 'member/my-member-connections.html'
-
-    def get_context_data(self, **kwargs):
-        """
-        Add connections and study_grants to the request context.
-        """
-        context = super(MyMemberConnectionsView, self).get_context_data(
-            **kwargs)
-
-        connections = [
-            item for item in self.request.user.member.connections.items()
-            if apps.get_app_config(item[0]).disconnectable
-        ]
-
-        context.update({
-            'connections': connections,
-            'study_grants': self.request.user.member.study_grants.all(),
-        })
-
-        return context
+from .mixins import SourcesContextMixin
+from .models import Member
 
 
 class SourceDataFilesDeleteView(PrivateMixin, DeleteView):
@@ -359,138 +67,6 @@ class DataSelfieFileDeleteView(PrivateMixin, DeleteView):
                                               user=self.request.user)
 
 
-class UserDeleteView(PrivateMixin, DeleteView):
-    """
-    Let the user delete their account.
-    """
-    context_object_name = 'user'
-    template_name = 'account/delete.html'
-    success_url = reverse_lazy('home')
-
-    def delete(self, request, *args, **kwargs):
-        response = super(UserDeleteView, self).delete(request, *args, **kwargs)
-
-        # Log the user out prior to deleting them so that they don't appear
-        # logged in when they're redirected to the homepage.
-        logout(request)
-
-        return response
-
-    def get_object(self, queryset=None):
-        return self.request.user
-
-
-class MyMemberConnectionDeleteView(PrivateMixin, TemplateView):
-    """
-    Let the user delete a connection.
-    """
-
-    template_name = 'member/my-member-connections-delete.html'
-
-    def get_access_tokens(self, connection):
-        connections = self.request.user.member.connections
-
-        if connection not in connections:
-            raise Http404()
-
-        access_tokens = AccessToken.objects.filter(
-            user=self.request.user,
-            application__name=connections[connection]['verbose_name'],
-            application__user__username='api-administrator')
-
-        return access_tokens
-
-    def get_context_data(self, **kwargs):
-        context = super(MyMemberConnectionDeleteView, self).get_context_data(
-            **kwargs)
-
-        connection_app = apps.get_app_config(kwargs.get('connection'))
-
-        context.update({
-            'connection_name': connection_app.verbose_name,
-        })
-
-        return context
-
-    def post(self, request, **kwargs):
-        connection = kwargs.get('connection')
-        connections = self.request.user.member.connections
-
-        if not connection or connection not in connections:
-            return HttpResponseRedirect(reverse('my-member-connections'))
-
-        if request.POST.get('remove_datafiles', 'off') == 'on':
-            DataFile.objects.filter(user=self.request.user,
-                                    source=connection).delete()
-
-        # TODO: Automatic list of all current studies.
-        if connection in ('american_gut', 'go_viral', 'pgp', 'wildlife'):
-            access_tokens = self.get_access_tokens(connection)
-            access_tokens.delete()
-
-            return HttpResponseRedirect(reverse('my-member-connections'))
-
-        if connection == 'runkeeper':
-            django_messages.error(
-                request,
-                ('Sorry, RunKeeper connections must currently be removed by '
-                 'visiting http://runkeeper.com/settings/apps'))
-
-            return HttpResponseRedirect(reverse('my-member-connections'))
-
-        if connection == 'twenty_three_and_me':
-            user_data = request.user.twenty_three_and_me
-            user_data.genome_file.delete(save=True)
-            django_messages.warning(
-                request,
-                ('We have deleted your original uploaded 23andMe file. You '
-                 'will need to remove your processed files separately on your '
-                 'research data management page.'))
-            return HttpResponseRedirect(reverse('my-member-connections'))
-
-
-class MyMemberStudyGrantDeleteView(PrivateMixin, TemplateView):
-    """
-    Let the user delete a study grant.
-    """
-
-    template_name = 'member/my-member-study-grants-delete.html'
-
-    @staticmethod
-    def get_study_grant(request, **kwargs):
-        try:
-            study_grant_pk = kwargs.get('study_grant')
-            study_grant = StudyGrant.objects.get(pk=study_grant_pk,
-                                                 member=request.user.member)
-        except StudyGrant.DoesNotExist:
-            return None
-
-        return study_grant
-
-    def get_context_data(self, **kwargs):
-        context = super(MyMemberStudyGrantDeleteView, self).get_context_data(
-            **kwargs)
-
-        study_grant = self.get_study_grant(self.request, **kwargs)
-
-        if study_grant:
-            context.update({
-                'study_title': study_grant.study.title,
-            })
-
-        return context
-
-    def post(self, request, **kwargs):
-        study_grant = self.get_study_grant(request, **kwargs)
-
-        if not study_grant:
-            return
-
-        study_grant.delete()
-
-        return HttpResponseRedirect(reverse('my-member-connections'))
-
-
 class ExceptionView(View):
     """
     Raises an exception for testing purposes.
@@ -498,52 +74,6 @@ class ExceptionView(View):
     @staticmethod
     def get(request):  # pylint: disable=unused-argument
         raise Exception('A test exception.')
-
-
-class MemberSignupView(AccountSignupView):
-    """
-    Creates a view for signing up for a Member account.
-
-    This is a subclass of accounts' SignupView using our form customizations,
-    including addition of a name field and a TOU confirmation checkbox.
-    """
-    form_class = MemberSignupForm
-
-    def create_account(self, form):
-        account = super(MemberSignupView, self).create_account(form)
-
-        # We only create Members from this view, which means that if a User has
-        # a Member then they've signed up to Open Humans and are a participant.
-        member = Member(user=account.user)
-
-        member.newsletter = form.data.get('newsletter', 'off') == 'on'
-        member.allow_user_messages = (
-            form.data.get('allow_user_messages', 'off') == 'on')
-
-        member.name = form.cleaned_data['name']
-
-        member.save()
-
-        # this may not be necessary, but we do return the account object here
-        # so if it's consumed by anything that uses the member we'll want it to
-        # have the up to date version that includes the user's preferences and
-        # name
-        account.user.member = member
-
-        return account
-
-    def generate_username(self, form):
-        """Override as StandardError instead of NotImplementedError."""
-        raise StandardError(
-            'Username must be supplied by form data.'
-        )
-
-
-class MemberLoginView(AccountLoginView):
-    """
-    A version of account's LoginView that requires the User to be a Researcher.
-    """
-    form_class = MemberLoginForm
 
 
 class OAuth2LoginView(TemplateView):
@@ -709,28 +239,6 @@ class AuthorizationView(OriginalAuthorizationView):
         return [self.template_name]
 
 
-class SourcesContextMixin(object):
-    """
-    A mixin for adding context for connection sources to a template.
-    """
-
-    def get_context_data(self, **kwargs):
-        context = super(SourcesContextMixin, self).get_context_data(**kwargs)
-
-        context.update({
-            'sources': {
-                'american_gut': UserDataAmericanGut,
-                'go_viral': UserDataGoViral,
-                'pgp': UserDataPgp,
-                'runkeeper': UserDataRunKeeper,
-                'twenty_three_and_me': UserDataTwentyThreeAndMe,
-                'wildlife': UserDataWildLife,
-            }
-        })
-
-        return context
-
-
 class ActivitiesView(NeverCacheMixin, SourcesContextMixin, TemplateView):
     """
     A simple TemplateView for the activities page that doesn't cache.
@@ -850,49 +358,3 @@ class HomeView(TemplateView):
             return redirect(settings.LOGIN_REDIRECT_URL)
         else:
             return super(HomeView, self).get(request, *args, **kwargs)
-
-
-class EmailUserView(PrivateMixin, DetailView, FormView):
-    """
-    A simple form view for allowing a user to email another user.
-    """
-    form_class = EmailUserForm
-    queryset = Member.enriched.all()
-    slug_field = 'user__username'
-    template_name = 'member/member-email.html'
-
-    def get_success_url(self):
-        return reverse('member-detail',
-                       kwargs={'slug': self.get_object().user.username})
-
-    def form_valid(self, form):
-        sender = self.request.user
-        receiver = self.get_object().user
-
-        if not receiver.member.allow_user_messages:
-            django_messages.error(self.request,
-                                  ('Sorry, {} does not accept user messages.'
-                                   .format(receiver.username)))
-
-            return HttpResponseRedirect(self.get_success_url())
-
-        form.send_mail(sender, receiver)
-
-        metadata = EmailMetadata(sender=sender, receiver=receiver)
-        metadata.save()
-
-        django_messages.success(self.request,
-                                ('Your message was sent to {}.'
-                                 .format(receiver.username)))
-
-        return super(EmailUserView, self).form_valid(form)
-
-    def get_context_data(self, **kwargs):
-        context = super(EmailUserView, self).get_context_data(**kwargs)
-
-        context.update({
-            'panel_width': 8,
-            'panel_offset': 2,
-        })
-
-        return context
